+++ src/resources/views/pages/home.blade.php
@@ -23,6 +23,7 @@
                         <li>Add UI to views for additional active column in Pricelists table (issue #78).</li>
                         <li>Rearrange menu to split Physical product store and Digital product store (issue #52).</li>
                         <li>Promotions, Products and Portfolios now allow translation (issue #9).</li>
+                        <li>Use Laravel authentication, removed Sentry dependency (issue #81).</li>
                         <li>More coming... (developing)</li>
                     </ol>
                     <h4>Bug fixes:</h4>